feat: Mooncake extension
