+++ Project.toml
@@ -34,11 +34,13 @@ TOML = "fa267f1f-6049-4f14-aa54-33bafae1ed76"
 [weakdeps]
 Enzyme = "7da242da-08ed-463a-9acd-ee780be4f1d9"
 JSON3 = "0f8b85d8-7281-11e9-16c2-39a750bddbf1"
+Mooncake = "da2b9cff-9c12-43a0-ae48-6db2b0edb7d6"
 SymbolicUtils = "d1185830-fcd6-423d-90d6-eec64667417b"
 
 [extensions]
 SymbolicRegressionEnzymeExt = "Enzyme"
 SymbolicRegressionJSON3Ext = "JSON3"
+SymbolicRegressionMooncakeExt = "Mooncake"
 SymbolicRegressionSymbolicUtilsExt = "SymbolicUtils"
 
 [compat]
@@ -59,6 +61,7 @@ Logging = "1"
 LossFunctions = "0.10, 0.11, 1"
 MLJModelInterface = "1.5 - 1.11"
 MacroTools = "0.4, 0.5"
+Mooncake = "0.4.124"
 Optim = "1.8 - 1.13"
 Pkg = "<0.0.1, 1"
 PrecompileTools = "1"
